新增(CoreLib): 支持命令直执行已注册 Action
- 为 CoreLib 管理命令补充 action/list/run 子命令，便于查看并直接触发当前 ActionRegistry 中的动作。

+++ EmakiAttribute/src/main/java/emaki/jiuwu/craft/attribute/service/AttributeTraceService.java
@@ -129,7 +129,7 @@ public final class AttributeTraceService {
             if (points <= 0) {
                 continue;
             }
-            addTrace(reconstructed, traces, parent.id(), points, "EmakiAttribute", "parent_attribute", parent.id(), parent.displayName() + " 加点", "", "", "parent_attribute", true, "points", filter);
+            addTrace(reconstructed, traces, parent.id(), (double) points, "EmakiAttribute", "parent_attribute", parent.id(), parent.displayName() + " 加点", "", "", "parent_attribute", true, "points", filter);
             for (Map.Entry<String, Double> bonus : parent.childBonuses().entrySet()) {
                 if (bonus.getValue() == null) {
                     continue;
